requirements.txt
streamlit
gtts
deep-translator
python-dotenv
openai
streamlit-lottie
requests
